requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]
python-dotenv==1.0.1
anthropic>=0.40.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
streamlit==1.39.0
pandas==2.2.3
python-multipart==0.0.12
requests==2.32.3
psycopg2-binary==2.9.10
alembic==1.13.3
